requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pony==0.7.10
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
Werkzeug==0.15.5
